Extract findUpSync into shared utility
Move the duplicated findUpSync function from 3 files into
packages/shared/findUp.ts and import it from there.
