+++ apps/backend/src/lib/appconfig.ts
@@ -1,8 +1,7 @@
 import { z } from 'zod'
-import path from 'path'
-import fs from 'fs'
 import dotenv from 'dotenv'
 import dotenvExpand from 'dotenv-expand'
+import { findUpSync } from '@shared/findUp'
 
 // Zod schema
 export const configSchema = z.object({
@@ -59,17 +58,6 @@ export const configSchema = z.object({
 
   DEEPL_API_KEY: z.string().optional(),
 })
-
-function findUpSync(name: string): string | undefined {
-  let dir = process.cwd()
-  while (true) {
-    const candidate = path.join(dir, name)
-    if (fs.existsSync(candidate)) return candidate
-    const parent = path.dirname(dir)
-    if (parent === dir) return undefined
-    dir = parent
-  }
-}
 
 if (!['production', 'staging'].includes(process.env.NODE_ENV!)) {
   // This will walk up from the current directory to find the first `.env` file
