+++ apps/frontend/vite.common.ts
@@ -1,8 +1,8 @@
 import path from 'path'
-import fs from 'fs'
 import os from 'os'
 import { loadEnv, type Plugin } from 'vite'
 import { getPackageVersion } from '../../packages/shared/version'
+import { findUpSync } from '../../packages/shared/findUp'
 
 export const hostname = os.hostname()
 export const mdnsName = hostname + '.local'
@@ -47,17 +47,6 @@ export const server = (mode: string, env: Record<string, string | undefined>, ap
         ],
       },
     },
-  }
-}
-
-function findUpSync(name: string): string | undefined {
-  let dir = process.cwd()
-  while (true) {
-    const candidate = path.join(dir, name)
-    if (fs.existsSync(candidate)) return candidate
-    const parent = path.dirname(dir)
-    if (parent === dir) return undefined
-    dir = parent
   }
 }
