+++ packages/shared/vite.common.ts
@@ -1,9 +1,9 @@
 import path from 'path'
-import fs from 'fs'
 import os from 'os'
 import { loadEnv, type Plugin } from 'vite'
 import mkcert from 'vite-plugin-mkcert'
 import { getPackageVersion } from './version'
+import { findUpSync } from './findUp'
 import { appConfigSchema } from './zod/config/appConfig.schema'
 
 export const hostname = os.hostname()
@@ -57,17 +57,6 @@ export const server = (mode: string, env: Record<string, string | undefined>, ap
         ],
       },
     },
-  }
-}
-
-function findUpSync(name: string): string | undefined {
-  let dir = process.cwd()
-  while (true) {
-    const candidate = path.join(dir, name)
-    if (fs.existsSync(candidate)) return candidate
-    const parent = path.dirname(dir)
-    if (parent === dir) return undefined
-    dir = parent
   }
 }
